+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+
+- Add some price method on `Rapa::Resources::ItemResource`
+
 ## v0.5.2
 
 - Support YYYY-mm format on `#publication_date` and `#release_date`